add custom hooks to index and about
