+++ src/pages/about.tsx
@@ -1,26 +1,46 @@
 import Head from 'next/head'
+import React from 'react'
 
 import { Footer } from '@/components/Footer/Footer'
 import { Header } from '@/components/Header/Header'
 import { Main } from '@/components/Main/Main'
+import { useBgLightBlue } from '@/hooks/useBgLightBlue'
+import { useCounter } from '@/hooks/useCounter'
+import { useInputArray } from '@/hooks/useInputArray'
 import styles from '@/styles/Home.module.css'
 
 import type { NextPage } from 'next'
 
-const About: NextPage = () => (
-  <div className={styles.container}>
-    <Head>
-      <title>Create App</title>
-      <meta name="description" content="Generated by create next app" />
-      <link rel="icon" href="/favicon.ico" />
-    </Head>
+const About: NextPage = () => {
+  const { count, handleClick, isShow } = useCounter()
+  const { text, array, handleChange, handleAdd } = useInputArray()
+  useBgLightBlue(count)
+  return (
+    <div className={styles.container}>
+      <Head>
+        <title>Create App</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
 
-    <Header />
+      <Header />
+      <div className={styles.container}>
+        <button onClick={handleClick}>ボタン</button>
+        <input type="text" value={text} onChange={handleChange}></input>
+        <button onClick={handleAdd}>追加</button>
+        {isShow ? <h1>{count}</h1> : null}
+        <ul>
+          {array.map(item => (
+            <li key={item}>{item}</li>
+          ))}
+        </ul>
+      </div>
 
-    <Main title="about" />
+      <Main title="about" />
 
-    <Footer />
-  </div>
-)
+      <Footer />
+    </div>
+  )
+}
 
 export default About
